feat: move cache & bull to global
