+++ packages/analytics/src/bootstrap/bootstrap.module.ts
@@ -1,11 +1,11 @@
-import { SeederModule, ServerAppModule } from '@metad/server-core'
 import { ServerAIModule } from '@metad/server-ai'
+import { provideBullModule, provideCacheModule, SeederModule, ServerAppModule } from '@metad/server-core'
 import { Logger, MiddlewareConsumer, Module, NestModule, OnApplicationShutdown } from '@nestjs/common'
 import { AnalyticsModule } from '../app.module'
 export * from '../ai/index'
 
 @Module({
-	imports: [ServerAppModule, ServerAIModule, AnalyticsModule, SeederModule]
+	imports: [provideCacheModule(), provideBullModule(), ServerAppModule, ServerAIModule, AnalyticsModule, SeederModule]
 })
 export class BootstrapModule implements NestModule, OnApplicationShutdown {
 	constructor() {}
