+++ packages/server/src/bootstrap/index.ts
@@ -139,3 +139,6 @@ export async function registerAllEntities(
 	// }
 	return allEntities;
 }
+
+export * from './cache'
+export * from './bull'
